reorganised engine directory
