+++ src/engine/input/Keyboard.hpp
@@ -11,9 +11,9 @@ class Keyboard {
 	// Returns the current state of all keys.
 	const std::array<KeyState, SDL_NUM_SCANCODES> &getKeyStates() const { return keyStates_; }
 
-	// Returns the state of a specific key. 
+	// Returns the state of a specific key.
 	const KeyState getKeyState(const SDL_Scancode key) const { return keyStates_[key]; }
-	
+
 	// Updates the key states based on the SDL event.
 	void update(const SDL_Event &event)
 	{
