+++ src/engine/types.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+// This file offers convenient access to all common types the engine provides.
+
+#include "input/KeyState.hpp"
+#include "types/ColorRGBA.hpp"
+#include "types/Rectf.hpp"
+#include "types/Recti.hpp"
+#include "types/Texture.hpp"
+#include "types/TextureFlip.hpp"
+#include "types/Vec2f.hpp"
+#include "types/Vec2i.hpp"
