+++ src/entities/sign.hpp
@@ -6,7 +6,7 @@
 #include "../components/Renderable.hpp"
 #include "../constants.hpp"
 #include "../ecs/ECSManager.hpp"
-#include "../engine/Vec2i.hpp"
+#include "../engine/types/Vec2i.hpp"
 #include <string>
 
 Entity instantiateSignEntity(ECSManager &ecs, Vec2i positionInTiles, const std::string &text = "")
@@ -14,7 +14,7 @@ Entity instantiateSignEntity(ECSManager &ecs, Vec2i positionInTiles, const std::
 	Entity npc = ecs.addEntity();
 
 	ecs.addComponent(npc, Positionable{{positionInTiles.x * TILE_SIZE, positionInTiles.y * TILE_SIZE}});
-	ecs.addComponent(npc, Renderable{{5*TILE_SIZE,1*TILE_SIZE},{TILE_SIZE, TILE_SIZE}, 0});
+	ecs.addComponent(npc, Renderable{{5 * TILE_SIZE, 1 * TILE_SIZE}, {TILE_SIZE, TILE_SIZE}, 0});
 	ecs.addComponent(npc, Collider{});
 	ecs.addComponent(npc, Interactable{text});
