+++ src/systems/DebugSystem.hpp
@@ -7,8 +7,8 @@
 #include "../ecs/ECSManager.hpp"
 #include "../ecs/Entity.hpp"
 #include "../engine/Engine.hpp"
-#include "../engine/Vec2f.hpp"
-#include "../engine/Vec2i.hpp"
+#include "../engine/types/Vec2f.hpp"
+#include "../engine/types/Vec2i.hpp"
 #include "../modules/Camera.hpp"
 #include "../modules/MapManager.hpp"
 #include "System.hpp"
