+++ src/ui/ConfirmationMenu.hpp
@@ -6,7 +6,7 @@
 #include "../modules/SaveGameManager.hpp"
 #include "../ecs/ECSManager.hpp"
 #include "../engine/Engine.hpp"
-#include "../engine/Vec2i.hpp"
+#include "../engine/types/Vec2i.hpp"
 #include "ListDialog.hpp"
 #include "TextDialog.hpp"
 #include "UIElement.hpp"
