+++ src/ui/MainMenu.hpp
@@ -3,8 +3,8 @@
 #include "../constants.hpp"
 #include "../ecs/ECSManager.hpp"
 #include "../engine/Engine.hpp"
-#include "../engine/Texture.hpp"
-#include "../engine/Vec2i.hpp"
+#include "../engine/types/Texture.hpp"
+#include "../engine/types/Vec2i.hpp"
 #include "../modules/GameStateManager.hpp"
 #include "../modules/MenuStack.hpp"
 #include "../modules/SaveGameManager.hpp"
